Move CompactionJobCommitterOrSendToLambda to compaction-core
